requirements.txt
asgiref==3.5.2
Django==4.1.2
django-ninja==0.19.1
pydantic==1.10.2
sqlparse==0.4.3
typing_extensions==4.4.0
